Added metrics framework and implementation (#910)
[feature] Added metrics framework and Prometheus implementation.
  - Enable by setting `metrics_provider_class` to `MetricsProviderPrometheus`.
  - Metrics endpoint can be found on http://localhost:9400/metrics by default.
  - port can be changed with `metrics_endpoint_port`. Set to 0 to disable http endpoint.


To gain insights into the state of running software, the software can expose metrics, which are then gathered by a metrics framework.
A popular framework for this is [Prometheus](https://prometheus.io/), but there are others.
This PR adds the ability for Moquette to expose metrics for such system to gather, and provides an implementation for Prometheus specifically.

By default, metrics are not gathered, and the system is only a shim that has minimal impact on performance. The Proteus implementation is a separate project that can be added to the classpath if desired. This minimises the impact for users that do not require metrics and does not change any dependencies for those use-cases.